A lot of changes to support the new Bean!
The new bean is called DrawingPanel and is basically the canvas. This bean has fires property change events that will let the program know when a drawing was added or removed. A further even will be added because alterations to the list can be made directly into the objects. Furthermore this canvas has a background color and a rotation.